Y3S1/DP/FINAL/Assigment/Assigment/src/main/java/accounts/AccountFactory.java
package accounts;


//refactor account using dp
public class AccountFactory {
    public static Account createAccount(Account.TYPE type, String accountNumber, double initialBalance) {
        switch (type) {
            case EUR:
                return new AccountEUR(accountNumber, initialBalance);
            case RON:
                return new AccountRON(accountNumber, initialBalance);
            default:
                throw new IllegalArgumentException("Unsupported account type: " + type);
        }
    }
}


